Include sources.list and apt-key in package
When the package contains the configuration of the repository, this eases
the installation and the upgrade to the package shipped by the Debian
project. On first install, only a download of the package file and install
with `apt install ./matrix-synapse_0.30.0-1_all.deb` is needed. When
Debian ships its official package an upgrade would remove the sources.list
and apt-key entry.

The key file was build with:

% curl 'https://matrix.org/packages/debian/repo-key.asc' \
  |gpg --homedir . --keyring temp-keyring.gpg --no-auto-check-trustdb \
  --trust-model always --ignore-time-conflict --no-options \
  --no-default-keyring --batch --yes --import
% gpg --homedir . --keyring temp-keyring.gpg --no-auto-check-trustdb \
  --trust-model always --ignore-time-conflict --no-options \
  --no-default-keyring --batch --yes --output debian/matrix-synapse.gpg \
  --export packages@matrix.org
% rm temp-keyring.gpg*